backend/requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
